Create js modules, which are imported and exported as required
