+++ index.css
@@ -207,7 +207,7 @@ footer {
   width: auto;
   align-items: start;
   border: 4px solid #000;
-  margin-bottom: 2vh !important;
+  margin-bottom: 1vh !important;
 }
 
 .footer-info {
