+++ index.js
@@ -1,12 +1,12 @@
 /* eslint-disable no-use-before-define */
 import { mainHeader } from './modules/header.js';
-import hideSome from './modules/pageHide.js';
+import pageSwitch from './modules/contentSwitch.js';
 import { mainFooter } from './modules/footer.js';
+import addNewBook from './modules/addbook-event.js';
 
 window.onload = () => {
   mainHeader.style.display = 'block';
-  hideSome();
+  pageSwitch();
   mainFooter.style.display = 'block';
+  addNewBook();
 };
-// const newBk = document.querySelector('#new-bk');
-// const currentDate = document.querySelector('.current-date');
