+++ modules/addbook-event.js
@@ -1,5 +1,9 @@
 // event to add book
-newBk.addEventListener('submit', (e) => {
+import { BooksClass } from './bookClass.js';
+
+export default function addNewBook() {
+  const newBk = document.querySelector('#new-bk');
+  newBk.addEventListener('submit', (e) => {
     e.preventDefault();
     const bookTitle = document.querySelector('#bk-title').value;
     const bookAuthor = document.querySelector('#bk-author').value;
@@ -8,3 +12,12 @@ newBk.addEventListener('submit', (e) => {
     awesomeBooks.ShowAllBooks();
     newBk.reset();
   });
+
+  const awesomeBooksIn = new BooksClass();
+  if (awesomeBooksIn.getLocalStorage().length > 0) {
+    awesomeBooksIn.ShowAllBooks();
+  } else {
+    const showBks = document.querySelector('#show-bks');
+    showBks.textContent = 'No books currrently added';
+  } // get store books or show message if empty
+}
