--- modules/getBooksStore.js
@@ -1,4 +1,0 @@
-const awesomeBooksIn = new BooksClass();
-if (awesomeBooksIn.getLocalStorage().length > 0) {
-  awesomeBooksIn.ShowAllBooks();
-}
