+++ modules/hideAll.js
@@ -2,8 +2,11 @@ import { addBookForm } from "./addBooksForm.js";
 import { collection } from "./collectionsDisplay.js";
 import { contact } from "./contactInfo.js";
 
-export default function hideSome() {
-    collection.style.display = 'block';
+export default function hideAll() {
+    collection.style.display = 'none';
     addBookForm.style.display = 'none';
     contact.style.display = 'none';
   }
+
+  
+
